third_party/WebKit/LayoutTests/http/tests/devtools/tracing/timeline-misc/timeline-event-causes-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Test that causes are correctly generated for various types of events.


Running: testTimerInstall
PASS - record contained Timer Installed
Promise @ setTimeoutFunction.js:

Running: testRequestAnimationFrame
PASS - record contained Animation Frame Requested
Promise @ requestAnimationFrameFunction.js:

Running: testStyleRecalc
PASS - record contained First Invalidated
styleRecalcFunction @ styleRecalcFunction.js:

Running: testLayout
PASS - record contained Layout Forced
layoutFunction @ layoutFunction.js:
PASS - record contained First Layout Invalidation
layoutFunction @ layoutFunction.js:
